Split blitz-headless crate out of blitz-test-harness
